+++ src/data/services/cacheService.ts
@@ -9,10 +9,17 @@ interface ICachedManifest {
     manifestData: Destiny2GameData;
 }
 
-const ManifestDatabaseName = "d2gunsmith_2_manifest_database";
+interface IPreferences {
+    language: ILanguageInfo;
+    hideRetiredPerks: boolean;
+    showCraftedBonus: boolean;
+    rawStatValues: boolean;
+}
+
+const ManifestDatabaseName = "d2weaponforge_manifest_database";
 const ManifestObjectStoreName = "manifestObjectStore";
-const ManifestCacheKey = "d2gunsmith_2_destiny_manifest";
-const LanguageCacheKey = "d2gunsmith_2_language";
+const ManifestCacheKey = "d2weaponforge_destiny_manifest";
+const PreferencesCacheKey = "d2weaponforge_preferences";
 
 export class CacheService {
     private readonly dbPromise: Promise<IDBDatabase>;
@@ -28,8 +35,8 @@ export class CacheService {
     public getCachedManifest = () => this.getValue<ICachedManifest>(ManifestCacheKey);
     public setCachedManifest = async (cachedManifest: ICachedManifest) => this.setValue(ManifestCacheKey, cachedManifest);
 
-    public getLanguage = async () => this.getValue<ILanguageInfo>(LanguageCacheKey);
-    public setLanguage = async (language: ILanguageInfo) => this.setValue(LanguageCacheKey, language);
+    public getPreferences = async () => this.getValue<IPreferences>(PreferencesCacheKey);
+    public setPreferences = async (preferences: IPreferences) => { console.log("setting preferences", preferences); this.setValue(PreferencesCacheKey, preferences); }
 
     private getValue = async <T>(key: IDBValidKey | IDBKeyRange): Promise<T | undefined> => {
         const objectStore = await this.getObjectStore("readonly");